requirements.txt
Babel==2.5.3
Django==2.0.1
django-phonenumber-field==2.0.0
phonenumberslite==8.8.10
pytz==2017.3
